free-llm-comparison/pyproject.toml
[project]
name = "llm-playground"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45",
    "openai>=1.0",
    "requests>=2.31",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = []
